refactor(endpoints): share in-memory limiters across integrations
